blogProject/requirements.txt
scrapy
dj-database-url==0.5.0
Django==2.0.5
django-widget-tweaks==1.4.2
Markdown==2.6.11
python-decouple==3.1
psycopg2==2.7.1
social-auth-app-django==2.1.0
django-allauth==0.35.0
djangorestframework
celery==4.1.0
flower==0.9.2
redis==2.10.6


